# swagger-api.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cocktail Ordering System API",
    "version": "1.0.0",
    "description": "A complete backend API for a cocktail ordering system with Express.js, Node.js, and MongoDB. Features JWT authentication, Paystack payment integration, and comprehensive admin management.",
    "contact": {
      "name": "Clyp AI",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
...
      "OrderItem": {
        "type": "object",
        "required": [
          "cocktail",
          "quantity",
          "price"
        ],
        "properties": {
          "cocktail": {
            "type": "string",
            "description": "Cocktail ID"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "description": "Quantity of the cocktail"
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "description": "Price of the cocktail at time of order"
          }
        }
      },
...
      "Order": {
        "type": "object",
        "required": [
          "customer",
          "items",
          "idempotencyKey"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "The auto-generated id of the order"
          },
          "orderNumber": {
            "type": "string",
            "description": "Auto-generated order number"
          },
          "idempotencyKey": {
            "type": "string",
            "description": "Unique key to prevent duplicate orders"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "subtotal": {
            "type": "number",
            "description": "Subtotal amount"
          },
          "totalAmount": {
            "type": "number",
            "description": "Total amount including any fees"
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "pending",
              "paid",
              "failed",
              "refunded"
            ],
            "description": "Payment status"
          },
          "fulfillmentStatus": {
            "type": "string",
            "enum": [
              "new",
              "preparing",
              "delivered",
              "cancelled"
            ],
            "description": "Order fulfillment status"
          },
          "paymentReference": {
            "type": "string",
            "description": "Payment reference from gateway"
          },
          "notes": {
            "type": "string",
            "maxLength": 500,
            "description": "Additional order notes"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/admin/orders": {
      "get": {
        "summary": "Get all orders with advanced filtering (Admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string"
            },
            "description": "Filter by customer state"
          },
          {
            "in": "query",
            "name": "paymentStatus",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "paid",
                "failed",
                "refunded"
              ]
            },
            "description": "Filter by payment status"
          },
          {
            "in": "query",
            "name": "fulfillmentStatus",
            "schema": {
              "type": "string",
              "enum": [
                "new",
                "preparing",
                "delivered",
                "cancelled"
              ]
            },
            "description": "Filter by fulfillment status"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter orders from this date (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter orders until this date (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search by order number or customer name"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of orders to return"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "orders": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin access required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/admin/orders/{id}/status": {
      "patch": {
        "summary": "Update order status with email notifications (Admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fulfillmentStatus"
                ],
                "properties": {
                  "fulfillmentStatus": {
                    "type": "string",
                    "enum": [
                      "new",
                      "preparing",
                      "in_route",
                      "delivered",
                      "cancelled"
                    ],
                    "description": "New fulfillment status"
                  },
                  "adminNote": {
                    "type": "string",
                    "maxLength": 500,
                    "description": "Admin note for the status update"
                  },
                  "sendEmailNotification": {
                    "type": "boolean",
                    "default": true,
                    "description": "Whether to send email notification to customer"
                  },
                  "estimatedDeliveryTime": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Estimated delivery time (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "order": {
                      "$ref": "#/components/schemas/Order"
                    },
                    "emailSent": {
                      "type": "boolean",
                      "description": "Whether email notification was sent"
                    },
                    "statusUpdate": {
                      "type": "object",
                      "properties": {
                        "previousStatus": {
                          "type": "string"
                        },
                        "newStatus": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/orders/track/phone/{phone}": {
      "get": {
        "summary": "Track orders by phone number (Public endpoint)",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "phone",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Customer phone number",
            "example": "+2348012345678"
          }
        ],
        "responses": {
          "200": {
            "description": "Orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "orders": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    },
                    "customerInfo": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "phone": {
                          "type": "string"
                        },
                        "totalOrders": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No orders found for this phone number"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}